chroe: rename package
and remove legacy files

Former-commit-id: 9a8aa5438cfa26b5797119447976ac98790083a8
